ARM: OMAP5: Add AXI error handler
In OMAP5 MPUSS, the internal CPU errors that occurs during the AXI bus
transactions are logged in the L2ECTLR register.

There are two types of errors that are logged.
- L2 internal asynchronous error caused by L2 RAM double-bit
  ECC error or illegal writes to the Interrupt Controller
  memory-map region.

- AXI asynchronous error caused due to illegal transactions
  with the local pheripherals.

These errors are enabled by default and communicated to the MPU via
a interrupt. There is no other data that is logged apart from the
above information. So register a interrupt handler to clear
these errors when they occur to avoid pending interrupts at GIC level.

Signed-off-by: R Sricharan <r.sricharan@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>