OMAPDSS: DSSCOMP: separate compositions for wb in m2m mode.
DSSComp can set separate compositions for WB M2M mode now.
The requirement for the active manager was removed because M2M
works with not active managers only.

Without this change the tiler will not able to provide the slots
to LCD manager. The blinking of panel happens because a composition
for LCD will not find a free slot. An enabling of the presentation
mode flag in DSS will give the posibility to split a big slots to
smaller ones and a LCD composition will able to find a free slot.

Change-Id: I04ea06351eb99c9849bd1cb4f8752e8b5ec46071
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>