OMAPDSS: DSS: Allow sending of compositions without WB buf
In some cases we have the situation when HWC has no buffer for WB but
a screen must be repainted. In this case a configuration passes
without a WB pipe and with disabled pipes to a second manager.

Now the checks for "WB enabled" removed. We can pass the compositions
without enabled pipes for second manager if this manager was
used by WB in prev compositions.

Change-Id: Ic4932b45682c88ae221d5a9c4f4f4e410d391198
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>