OMAPDSS: DSSCOMP: Allow multiple comps without WB buffer.
In some cases (like reconfiguring of resolution, etc) HWC has no
buffer for WB but should to repaint a screen. In this case HWC
passes the composition without WB buffer and with disabled pipes
for a second manager.

The check was changed to allow posting of multiple compositions
without WB to second manager which is inactive.

Change-Id: I8a4fcd0323d766605b829b27c7c2ce8695c41512
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>