wlcore: correctly check state before regdomain conf
The wlcore state was checked without the mutex being taken. This leads
to WARN_ONs sometimes if a notification arrives when the driver
is on, but the mutex is only taken after it is off. This usually
happens if stopping the driver while connected to a network.

Signed-off-by: Arik Nemtsov <arik@wizery.com>
Signed-off-by: Luciano Coelho <coelho@ti.com>
(cherry picked from commit 75592be5e35b31eb28dacf578cfe82f0e9bc0ac0)

Signed-off-by: Felipe Balbi <balbi@ti.com>