omap: remoteproc: adjust CMA base addresses to be within 64MB
The remoteproc CMA based addresses have been adjusted to lie
within the first 64MB of DDR, and start exactly at 0x82000000.

The CMA regions will be included as part of the kernel memory
and should lie outside of the memory that would be used for
vmalloc. This adjustment allows flexibility of supporting
minimum footprint ducati for 256M to 512M memory
configurations  of Android with vmalloc size
varying between 128M to 240M.

Change-Id: I7d3353c438cfd8551c07940072dc4b7ff0eb1e50
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Sundar Raman <a0393242@ti.com>