ARM/dts: omap5-panda: Remove lpddr2 device node
The memory part connected to OMAP5 panda board is DDR3.
Since both DVFS and temperature alert are not supported with
DDR3, do not register the device. So add a new omap5 panda
dts file without a ddr3 device node.

This fixes an unstability issue, which is otherwise seen
when EMIF driver is enabled on panda board.

Signed-off-by: Sricharan R <r.sricharan@ti.com>