ARM/dts: omap5-uevm: Remove lpddr2 device node
The memory part connected to OMAP5 uevm board is DDR3.
Since both DVFS and temperature alert are not supported with
DDR3, do not register the device. So add a new omap5 uevm
dts file without a ddr3 device node.

This fixes an instability issue, which is otherwise seen
when EMIF driver is enabled on uevm board.

while here, also rename the omap5-evm.dts to omap5-sevm.dts
to avoid confusion.

Signed-off-by: Sricharan R <r.sricharan@ti.com>
[rnayak: renamed evm to sevm, minor changlog fixes]
Signed-off-by: Rajendra Nayak <rnayak@ti.com>