remoteproc: match the secure parameters to OMAP4 ION carveout values
The secure decoded buffer addresses are currently hard-coded in the
secure parameters, and these need to match the corresponding ION
carveout set-aside for the 2D TILER heap. The decode_buffer_address
and decoded_buffer_size are corrected to match the corresponding
values for OMAP4.

TODO: This is a quick patch that solves the secure playback firewall
issues for OMAP4. The current hardcoding code is a hack by itself,
and needs to be done properly if it were to scale for OMAP5 as well.

Change-Id: I525f214f7c66800995fa8310f9e4a0fe3be32d7b
Signed-off-by: Suman Anna <s-anna@ti.com>