ARM: dts: AM33XX: Fix gpio numbering to match hardware/TRM
With DT support, where naming convention is based on base-addr
and not id, so we should follow TRM/Spec numbering label.

This patch changes GPIO numbering as per TRM, as gpio0-3.

Matt Porter had submitted base patch sometime back but it
didn't make it to the mainline -
https://patchwork.kernel.org/patch/1433001/

Later DT nodes for gpio based keypad and led driver got
merged, so this fix needs to propagate to board dts files
as well.

Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>
Signed-off-by: Matt Porter <mporter@ti.com>
Acked-by: Peter Korsgaard <jacmet@sunsite.dk>
Cc: Benoit Cousson <b-cousson@ti.com>