ARM: OMAP: dpll: enable bypass clock only when attempting dpll bypass
.set_rate for non core DPLLs attempts an enable of bypass clk as well
as ref clk for every .set_rate attempt. Early at boot, while some of
these DPLLs are locked, this causes an ordering issue.

For instance, on OMAP5, the USB DPLL derives its bypass clk from ABE DPLL.
If the USB DPLL is attempted to be locked before the ABE, the enable of
USB bypass clk (derived from ABE DPLL) inturn causes the ABE DPLL to be
locked and fails as the M,N values are yet to be programmed.

To get rid of this ordering needs, enable bypass clk for a DPLL as part
of its .set_rate only when its being put in bypass.

Reported-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>