iommu/omap: add support for secure mode
The omap iommu module has been enhanced to provide the support
for starting a remote processor in a mode suitable for running
secure applications.

The mode and ttbr value to be used in secure mode are configured
through the domain-specific iommu_data. The iommu_data needs to
be set before attaching the iommu device. In secure mode, all
iommu operations to program a PTE entry or TLB entry are blocked.
The TTB is configured to a pre-determined L1 table which is
firewalled, and the sanity of TTB is continuously checked in
the secure code.

Change-Id: I0149016502602cb3ce69486a49e6df8d1c5ca645
Signed-off-by: Suman Anna <s-anna@ti.com>