remoteproc: configure iommu for secure and normal modes
The remoteproc driver is adapted to retrieve and add the
appropriate mode and data information to the iommu domain,
as per its secure state. This information will be used by
the iommu driver during its init, and block any API that
might tend to program the L2MMU at runtime in secure mode.

Change-Id: I1887f3eb1616f9b74a9b95b3915bf66e02ee75d4
Signed-off-by: Suman Anna <s-anna@ti.com>