remoteproc: program the iommu after handling resources
The remoteproc driver currently maps the different processor
memories while processing the various entries in the resource
table. This logic has been reworked to facilitate the iommu
programming model for secure mode.

The resources need to be handled first to gather the memory
translation data that is required for parsing the iommu data
for the secure mode. The iommu can then be configured based
on the mode the remoteproc is being booted in.

Change-Id: I0ffd8ed6a6fadf0f9fb0492b2832d1ff3353936c
Signed-off-by: Suman Anna <s-anna@ti.com>