remoteproc: split the secure sections parsing and authentication
The functionality in the rproc_secure_boot code is split into
two separate functions, one for parsing the secure data sections,
and the other for performing the image validation and configuring
the hardware for secure mode.

This is done so as to retrieve the ttbr value earlier than the
authentication step for configuring the iommu driver properly.

Change-Id: I53d131d79e2230c1b479eb6bf39e182f35c65ffc
Signed-off-by: Suman Anna <s-anna@ti.com>