Revert "sgx_omaplfb: cleanup static configuration"
This reverts commit a33f0a13e374e7d3aa94678ec7c818f190a4bab4.

The configuration of framebuffers was moved to board files and
The default configuration was removed from OmapLFB after it.
This brokes Blaze because Blaze has no specific configuration
for OMAPLFB driver.

This revert adds a default configuration for devices which
has no a specific configuration for OmapLFB in a board file.

Change-Id: I8b5b0f0e21c3cafac43a528499c1cd156d5c3ef4
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>