ARM: OMAP: remoteproc: add carveout memory pool configuration
The carveout memory pool data information is added to the OMAP
remoteproc platform data. A new configurable menuconfig option
has been added to help choose between the carveout pool or the
rproc device-specific CMA pool. The default option is left to
use the CMA pool.

The carveout memory for remoteproc will still be reserved at board
initialization time, but the allocation mechanism is dictated by
the configuration option. The option controls the configuration
choice for all the remote processors on the SoC, and cannot be
enabled differently for different processors.

Change-Id: Ic14734fbc3d64a9ee1ebf7c4933f588e76f9c8bc
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>