OMAPDSS : DSSCOMP : Avoid pushing content if no panel is active
Currently we push content on a blanked manager in case it is
default HDMI display. This causes a lot of logging in dsscomp call
back when we apply blank to manager.

Check added to see if number of overlay in the composition are not
zero.

In HWC we are checking if the default display is not present
we will populate the composition number of overlay as zero.

Dependencies:
This patch has dependency on HWC patch mentioned below..
http://review.omapzoom.org/#/c/26472/ - JB release

Change-Id: Icadfecb8f88adf3c6020b17d1869b5bb49b0efe6
Signed-off-by: Arthur Philpott <arthur.philpott@ti.com>