remoteproc: add .get & .set virtio config ops implementations
The virtio config ops functions for get and set are added through
rproc_virtio_get and rproc_virtio_set functions. These functions
are used in a non-standard way and are overloaded primarily to allow
the virtio rpmsg bus infrastructure to query the remoteproc driver
about it carveout pool configuration and to allocate/free the vring
buffers.

Change-Id: I6fbd6d1cb502b34f556469a5238b4b18128c2171
Signed-off-by: Suman Anna <s-anna@ti.com>