rpmsg: add support for allocating vring buffers from carveout
The current rpmsg code relies and uses the same CMA pool associated
with a rproc device. Support has been added to allocate the vring
buffers from a remoteproc's carveout pool. The virtio rpmsg bus
infrastructure does not have direct access to the remoteproc pool,
so it is extended using the virtio config ops. The remoteproc
carveout configuration is also retrieved to help with the proper
cleanup of the buffers.

The remoteproc carveout allocation returns both the physical address
and a virtual address. The virtual address is an ioremap'd address
and is not exactly the equivalent to the virtual address retrieved
from a dma_alloc_coherent call. Logic is adjusted to make the
underlying virtio publish a direct-mapped kernel address so that
the offsets are identical between the virtual and physical addresses.
This is essential for the remote processor to perform the address
translation properly using address masking.

Change-Id: I55848e9f5f3a4429f968285cddf3efa1feb6b47a
Signed-off-by: Suman Anna <s-anna@ti.com>