remoteproc: add support to use a carveout memory pool
The remoteproc driver relies on device-specific CMA pool for
allocating memory for various firmware image sections. Support
has been added now to let the remoteproc driver choose to use a
carveout pool instead of the CMA pool.

This option is added to allow flexibility and robustness for
allocations within the remoteproc driver. The carveout memory
option is especially beneficial for improving the memory allocation
times and for avoiding any possible remoteproc failures due to
run-time failure of a CMA allocation. This option provides this
reliability at the expense of a static carveout.

The remoteproc driver is self-configured to use a carveout pool
or the CMA pool based on whether a carveout pool data is supplied
or not during the rproc registration. The carveout pool data will
need to be constructed during the remoteproc device creation.

NOTE: The current changes customize the carveout allocations
specifically for the OMAP firmware sections because of its
hardcoded adherence to the CMA assignments, specifically for
the vrings and vring data buffers.

Change-Id: I396314d4e484e3e9f33f4fbf9023764a4ab9684f
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>