remoteproc: disable auto-suspend when entering secure mode
Secure module reads iommu ttbr register in order to make sure that the
iommu is using a firewalled MMU table. The clocks to iommu therefore
cannot be cut during secure play back. The remoteproc auto-suspend
is disabled for ipu processor when entering secure mode, and is
reenabled when switching back to normal mode.

Change-Id: I64b0b3acc7d9526b5b538104f7946d0e8ed0a537
Signed-off-by: Suman Anna <s-anna@ti.com>