omapdss: Add a DPI driver for DRA7xx SoC
Create a DPI driver for DRA7xx DSS, the only difference it has compared to the
OMAP DPI driver is that it uses dss core DPLLs as an alternative clock source
and not the DSI PLLs.

These drivers can be merged later on when DSI PLL code is made more generic and
is moved to dss core.

Signed-off-by: Archit Taneja <archit@ti.com>