omapdss: Add dss dplls to dss family
DRA7x has video DPLLs similar to that of the DSI PLLs on OMAP. However, they
aren't a part of DSI, and belong to the DSS core/ DSS family of registers. Tie
the video DPLLs to the dss core platform device.

Borrow the code from DSI to power on, configure and lock the PLLs. The apis
provided will be used by the DRA7x DPI driver to get a desired pixel clock.

Signed-off-by: Archit Taneja <archit@ti.com>