omap: watchdog: Use devm_request_irq instead of request_irq()
The missing free_irq() as part of the driver removal,
for the request_irq() done when the device is probed
results in a crash during the driver/module removal.

[  567.951568] Unable to handle kernel paging request at virtual address bf005088
[  567.959106] pgd = c0004000
[  567.961914] [bf005088] *pgd=aecc8811, *pte=00000000, *ppte=00000000
[  567.968475] Internal error: Oops: 80000007 [#1] SMP ARM
[  567.973907] Modules linked in: [last unloaded: omap_wdt]
[  567.979431] CPU: 0    Not tainted  (3.8.4-00186-ga9a396d #1)
[  567.985321] PC is at 0xbf005088
[  567.988616] LR is at handle_irq_event_percpu+0x64/0x244
[  567.994049] pc : [<bf005088>]    lr : [<c00a9b84>]    psr: 60000193
[  567.994049] sp : c078ded0  ip : 00000000  fp : c082a358
[  568.005981] r10: 00000000  r9 : c082a344  r8 : 00000000
[  568.011413] r7 : 00000070  r6 : 00000070  r5 : ee80aa10  r4 : eee2ba40
[  568.018188] r3 : 00000000  r2 : bf005088  r1 : eed18750  r0 : 00000070
[  568.024993] Flags: nZCv  IRQs off  FIQs on  Mode SVC_32  ISA ARM  Segment kernel
[...]

Use devm_request_irq() instead of request_irq(), so the
free_irq() is automatically handled during the module removal.

And also avoiding an unnecessary check.
irq field in omap_wdt_dev is used to check if irq is registerd or not.
Since driver probe comes out when request_irq fails, the above field is
not needed. Removing the same.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>