OMAPDSS: DSSCOMP: fix blank composition at suspend
mgrs array in composition has to be filled sequentially with mgr
data, while in HDMI as primary display scenario mgr1 is actually
placed in second position in array. It confuses gralloc_queue()
and it fails to correctly blank screen. This fix aims to handle
this case both with legacy and presentation cases.

Change-Id: I1a45dfa7f7e8ce63e4706f743bc1db2a93e5b8b2
Signed-off-by: Sergiy Kibrik <sergiy.kibrik@globallogic.com>