ASoC: omap-abe-twl6040: Add VAUX2 regulator for AV switch
An AV switch is used in Blaze/SDP4430 to switch headset jack pin 3
between headset microphone and video output. AV switch is then a
consumer of VAUX2 supply. The corresponding consumer-supply is added
to have VAUX2 supply when headset microphone is in use.

This is considered non-essential, so if the driver is unable to
acquire the VAUX2 supply, it will gracefully continue.

Change-Id: I749756a3b691369d696fcf88630936f8354170ba
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>
Signed-off-by: Victor Kleinik <x0150886@ti.com>