ARM: OMAP5: Cpufreq: Enable cpufreq for OMAP5
The patch enables DVFS for OMAP5 platform using the cpufreq-cpu0 driver.
The smps123 regulator and the functional clock for MPU are hooked to the
cpu0 device. This patch uses i2c1 and not the i2c_sr.

The voltage and frequency values work well only on NOM samples
and are supposed to work properly only with ABB/AVS.

TODO: Add SPEEDBIN OPP after ABB and AVS support so the cpufreq works
on all samples seamlessly.

Signed-off-by: J Keerthy <j-keerthy@ti.com>
Signed-off-by: Tero Kristo <t-kristo@ti.com>
Acked-by: Nishanth Menon <nm@ti.com>