rpmsg: omx: fix a race condition in rpmsg_omx_probe()
The rpmsg_omx is a client driver on the rpmsg bus, providing a user
space interface to the remote OMX server (represented as a rpmsg_omx
device). The rpmsg_omx device(s) are created at runtime when a remote
processor publishes the device(s) through the rpmsg bus's name service
announcement. The devices are destroyed during a remoteproc recovery
process, unless there are open references to the device. Such pre-
existing handles are rechecked for during the driver probe.

The current rpmsg_omx probe is racy when checking for a pre-existing
omx service device. The rpmsg omx devices are created by a kthread
(the vq messages in omap_remoteproc are processed through a kthread),
and so the rpmsg_omx_probe can be executed in parallel in a SMP
environment. An error can occur in _match_omx_service cb ("Unable to
handle kernel NULL pointer dereference at virtual address 0000002c")
within the probe due to a NULL dev pointer. This race condition is
possible any time a remote processor is booted up, including during
the kernel boot.

The race condition scenario is explained below:

thread1:                          thread2:
...                               ...
rpmsg_omx_probe() for 1st device  ...
  omxserv=idr_for_each(...)       ...
    _match_omx_service(...)       ...
  ...                             rpmsg_omx_probe() for 2st device
  idr_get_new(...) _               omxserv = idr_for_each(...)
  ...               |                _match_omx_service(...)
                    |
  cdev_init(...)    |_               Oops! omxserv->dev is NULL for 1st device
  cdev_add(...)     |
  ...              _|
  omxserv->dev=device_create(...)
  ...

The race condition can be eliminated by making sure the service
channel is added to the list only after the device node is created,
but protecting the entire probe using a lock ensures that the device
name check is also atomic. The spinlock defining the lock is converted
to the more suitable mutex.

Change-Id: Ifcc2e505da6ae32fcb67303852140c98f4200214
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>
[s-anna@ti.com: revise patch description and retain variable name]
Signed-off-by: Suman Anna <s-anna@ti.com>