HACK: ARM: OMAP5: Deny EMIF idle is DDR3 is in use
DDR3 low power transition state is under debug at the moment.
Hence, as a temporary WA, set EMIF not no idle state inorder
to prevent DDR3 corruption in low power transition.

NOTE: as a result of this patch core power domain will not idle
any longer either.

As part of this change - remove the shadowed variables for emif_clkdm
and mpuss_clkdm which are static to the function but shadowed in static
dependency lookup.

Change-Id: I6d8413f65579a08b0bf825bb06936b4563b74977
Signed-off-by: Nishanth Menon <nm@ti.com>