omapdss: displays: Add tfcs9700 DPI panel driver
TFCS9700 is a 800x480 DPI LCD panel found on the Vayu display daughter card.
The card also contains an i2c brightness controller/gpio expander(TLCxxxx) which
drives a pwm signal for the panel and also provides some gpios to configure some
of the panel pins.

Add the DPI panel driver as an omapdss driver, and add the i2c controller as a
client driver in the same file for now. In the longer run, we would have a
separate mfd driver for the TLC chip.

Signed-off-by: Archit Taneja <archit@ti.com>