remoteproc: use kernel memory to communicate secure parameters
Use a kernel allocated memory buffer to communicate the contents
of secure_params structure with the TEE code. It was observed
that when using ioremap buffers, the SMC driver resolved an
address that did not correspond to the desired memory location.
Thus, when TEE code attempted to access the memory, it would
trigger a DATA_ABORT when it attempts to access invalid memory.

Change-Id: I43e12dad2aac3006983297bab223303cd8957bcd
Signed-off-by: Carlos Leija <cileija@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>