TI-Integration:Fix auto merge issue on 33xx hwmod for ocmc ram
The auto merge of the omap_hwmod_33xx_data.c seemed to have removed
entries for the ocmram initilization.

The below commit seems to have been over written in a merge.

commit e70216ee5b50ac2d0015c08701e193bb2da983d2
ARM: OMAP2+: AM33XX: hwmod: Register OCMC RAM hwmod

Did not seem to be complete in the merge.
So adding back the code to register the OCMC RAM hwmod

Signed-off-by: Dan Murphy <dmurphy@ti.com>