OMAPDSS: HDMI: Add level shifter control and state
This patch adds a level shifter mechanism that provides control of the
CT_CP_HPD and LS_OE pins on the level shifter.

There are four different possibilities for the outputs of the pins,
but only 3 are valid given that the level shifter is disabled completely
when CT_CP_HPD is low.

The following are the states for the level shifter:
LS_DISABLED: both CT_CP_HPD and LS_OE are off.  The chip is off.
LS_HPD_ON: CT_CP_HDP is on, LS_OE off.  5V rail and HPD detection is on.
LS_ENABLED: CT_CP_HPD and LS_OE are on.  Level shifters are active.

The hdmi_set_ls_state is a helper function that allows for the correct
state of the CT_CP_HPD and LS_OE pins to be set given the desired state.

Signed-off-by: Andy Gross <andy.gross@ti.com>