omapdss: hdmi: select I2C_ALGOBIT in Kconfig
The hdmi driver uses i2c bitbanging to read edid for OMAP5. The core hdmi
driver(hdmi.c) adds an i2c adpater based on the i2c bitbanging driver available
in the kernel.

This CONFIG should be selected only for OMAP5 HDMI, but since the adapter is
added in hdmi.c, it's select for both OMAP4 and OMAP5 HDMI for now. This can
be clenaed up later.

This prevents a build break when DRM is not selected, DRM core selects
I2C_ALGOBIT by default, so the compile error isn't noticed when omapdrm is
built.

Signed-off-by: Archit Taneja <archit@ti.com>