omapdss: hdmi: fix deepcolor mode configuration
The hdmi driver supports deep color modes. However, there is a specific check
which doesn't allow 1080p timings or higher to be shown 36 bit deep color mode.

The more generic check would be to check if the pixel clock in a deep color mode
is less than the maximum pixel clock the TV overlay manager can support for the
given OMAP. Make this change.

Also, make sure that hdmi.dssdev is assigned by the end of probe, this param is
needed by the deepcolor sysfs attribute file.

Signed-off-by: Archit Taneja <archit@ti.com>