omap5: panda: initialize vram via android-display
This patch adds logic to initialize vram via
android-display similar to OMAP4 Tablet and OMAP5 sEVM
so that the vram requirements are calculated using
the display size accurately rather than being hardcoded.

Also, the number of vram buffers have been increased
from 2 to 4 inorder to support blitter functionality
on secondary display.

Inorder to fully enable blitter, there is a corresponding
filesystem change in device/ti/panda5 that is required
to update the bootargs to not pass the vram sizes which
will override the android-display setting.

Change-Id: I490c22f154fac08eaaf481831fec135915253c1f
Signed-off-by: Sundar Raman <a0393242@ti.com>