remoteproc: check for proper rproc state before power off
Add a check for the current rproc state in rproc_shutdown() to prevent
powering off the remote processor if it has not been booted completely.
The remote processor would remain in offline state if the boot fails
for any reason, and the cleanup would be handled automatically in the
rproc_fw_boot() function upon a boot failure.

Change-Id: Ie01eee861c0a75d134bda55f573eb42a85ff8b81
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>