ARM: OMAP: iommu: do shutdown instead of idle on runtime suspend
This patch fixes an issue with resuming/booting the remote processors
in OMAP4+ SoCs when the particular domain goes through a retention
(CSWR or OSWR) to active state transition.

The Attribute MMU (AMMU) registers are not accessible from the MPU
side, and are programmed/restored by the SYS/BIOS RTOS running on the
remote processors during one of its initial steps. The remote cores
expect the Unicache to be in disabled/bypass state before this initial
programming step, in both the cold and warm boot sequences. The AMMU/
Unicache module logic is controlled by the specific h/w reset line
associated with the MMU. Any retention to active transition of the
corresponding remote processor subsystem's power domain, with this
h/w reset line deasserted, applies a retention reset and causes the
AMMU to lose context, but with Unicache/AMMU module enabled. This
hangs the remote processor upon boot/resume as an internal bus fault
is thrown due to the lack of valid AMMU programming registers for the
instruction space that the processor is executing. This scenario is
applicable both during a warm boot (resume) and a cold boot (if there
is sufficient time between the IOMMU programming and the processor
reset deassertion for the domain to enter a retention state).

The IOMMU power management including Device OFF support is currently
implemented through the runtime pm framework. The runtime suspend for
the IOMMU driver is invoked whenever the remote processor is put into
a suspend state. The IOMMU is idled at present on runtime suspend with
the associated h/w reset line left in deasserted state. This leaves
the particular sub-modules clocked and enabled, and susceptible to
the above AMMU context loss problem.

The issue is fixed by defining an IOMMU specific omap_device_pm_latency
structure and using a omap_device_shutdown_hwmods() instead of the
default omap_device_idle_hwmods() within it for deactivate function.
This ensures that the OMAP hwmod framework will assert the h/w reset
line upon a runtime suspend and allows the IOMMU to be programmed /
restored back properly upon runtime resume. Logic (implemented through
dev_pm_qos latency constraints) is already present in the iommu driver
preventing any OSWR/CSWR transitions on the corresponding power domain
during active use-cases.

Change-Id: I2d35d35129e313620ba86e6ed6ce6a605f056942
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
[s-anna@ti.com: rewrite commit description completely]
Signed-off-by: Suman Anna <s-anna@ti.com>