ARM: OMAP3+: DVFS: Remove unnecessary voltage ceiling
_dep_scan_table() ceil dependent voltage to a valid OPP of "primary" device.
It was needed only in case we have frequency request to
keep voltage and frequency in sync.

Now frequency requests are dropped and voltage
ceiling is not needed. It even became harmful.
For example in OMAP4470 High OPP set L3 doesn't have a real OPP100.
So OPP100 voltage dependency request to CORE
will be always ceiled to OPP119 voltage.

Just drop this code.

Change-Id: I51b8e43b695292fdc78ed07a15c529f52b9c2ba5
Signed-off-by: Taras Kondratiuk <taras@ti.com>