thermal: consider emul_temperature while computing trend
In case emulated temperature is in use, using the trend
provided by driver layer can lead to bogus situation.
In this case, debugger user would set a temperature value,
but the trend would be from driver computation.

To avoid this situation, this patch changes the get_tz_trend()
to consider the emulated temperature whenever that is in use.

Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>
[Backported to 3.8]
Signed-off-by: J Keerthy <j-keerthy@ti.com>