OMAPDSS:HDMI: Change PLL calculations
This patch modifies the algorithms used to determine the correct PLL
settings for the HDMI TMDS output.  Instead of using fixed REGM2 values
we search for a valid combination of REGN, REGM, REGM2 that satisfy the
constraints for the requested frequency.

The computation involves looping through the values of REGN and trying to
find an adequate REGM and REGM2 (if applicable).  This results in the lowest
value for REGN that will satisfy the frequency request.

This patch also modifies the API used for the hdmi_compute_pll function.
If a valid PLL combination cannot be found, we return an error back to the
caller.

Signed-off-by: Andy Gross <andy.gross@ti.com>
Signed-off-by: Archit Taneja <archit@ti.com>