ARM: OMAP5: hwmod data: set init_no_reset flag for ipu iommu
This patch fixes a problem with the IPU boot sequence caused
by an improper AMMU/Unicache state upon a retention->active
transition of CORE domain. The AMMU/Unicache is enabled by the
SCACHE_CONFIG.BYPASS bit - a value of 1 enables the cache and
mandates all addresses accessed by M3 be defined in the AMMU.
This bit is not programmable from the A9-side. The M3 boot
sequence starts out with the AMMU/Unicache in disabled state,
and SYS/BIOS programs the AMMU regions and enables the Unicache
during one of its initial boot steps. Out of retention, the
SCACHE_CONFIG.BYPASS bit is enabled by default, irrespective of
whether it was previously enabled or not.

By not setting this flag in the hwmod database, the hwmod
initialization enables the clocks and deasserts RST3. And when
the CORE domain hits either of OSWR or CSWR, the BYPASS bit gets
auto-set, and the M3 boot throws an exception as soon as the M3
core's reset is deasserted since the boot address is typically
not defined in any valid AMMU regions. Setting this flag allows
an IPC driver to take out the resets as required, after doing the
necessary programming. This also keeps the entire sub-system in
RESET if no IPC driver is selected in the kernel.

Change-Id: Ifc916db88247d2de9025974eb595fdc32b0e9669
Signed-off-by: Suman Anna <s-anna@ti.com>