thermal: bandgap: Meld TSHUT config functions into one function
On OMAP4470 CONTROL_TSHUT_THRESHOLD was changed to one time writable
register, in difference to OMAP4460 where this register could be
rewrited at any time. Because of this we need to configure it using
one write operation.
So considering this, high and low thresholds should be configured
in one function. To accomplish that changing temp_sensor_configure_tshut_thigh()
and temp_sensor_configure_tlow() to just temp_sensor_configure_tshut().

Change-Id: Ib41b77f83cf07bd13c5a880dbd7cc652bdf73e77
Signed-off-by: Ruslan Ruslichenko <ruslan.ruslichenko@ti.com>