thermal: bandgap: Add TSHUT_CONFIG_ONCE feature
As on OMAP4470 TSHUT config is one time writebale register, we need to ensure
that driver does not try to reconfigure it after it was configured earlier during
boot.
For OMAP4470 GP devices content of tshut config register is not restored
after OFF mode, and it need to be configured manually to prevent spurious TSHUT
interrupt generation.

Change-Id: I6f232d02b534926fd987aaa4b4a87aa522cc8079
Signed-off-by: Ruslan Ruslichenko <ruslan.ruslichenko@ti.com>