mailbox/omap: check for status of platform_get_irq
The IRQs have to be defined in both the hwmod file as well
as the DT node (depending on a non-DT or DT boot). The
status of the platform_get_irq function therefore needs to
be checked to make sure the irqs are defined for the mailbox
platform device for the probe to be successful.

Signed-off-by: Suman Anna <s-anna@ti.com>