ARM: dts: DRA7: Add mailbox nodes
DRA7xx has 13 system mailboxes, add the DT nodes for
all these 13 mailboxes. The unused mailboxes have been
disabled, and only mailboxes 1, 5 and 6 are enabled.
The remaining mailboxes can be enabled when a usecase
arises and their irq/xbar numbers are identified.

Currently, the FIFO assignment and target processor
assignment data is added only for mailboxes 1, 5 and 6.
The data is chosen to match the current mailbox assignments
in IPC 3.x with MPU receiving interrupts using user id
2 on mailboxes 5 & 6, and user id 0 on mailbox 0.

NOTE: The DT node data uses specific interrupt numbers
for mailboxes 5 & 6 and the actual interrupts are raised
only when the corresponding crossbar registers are
programmed correctly.

Signed-off-by: G Anthony <a0783926@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>