ARM: omap4plus_pm_min_defconfig: Enable Thermal support
It is not safe to work without thermal framework support, as device
can get overheat and corrupted. Additionally some critical thermal
registers need to be configured (like TSHUT registers) or problem
with OFF mode, warm reset may appear or device even can be damaged
if TSHUT interrupt is not configured and handled properly.

Change-Id: Ia4775e5f5d7411b995fbe76d3999a9a38c0d496a
Signed-off-by: Ruslan Ruslichenko <ruslan.ruslichenko@ti.com>