OMAPDSS: OMAPFB : Add support for default display scaling.
Currently it is assumed that frame buffer size
is same as default panel size. With this assumption
GFX pipeline is always presented to default display FBDev.
GFX pipeline cannot scale so we cannot have a frame
buffer which is not equal to panel size (LCD | HDMI).
Updated the current logic of omapfb to pick video pipelines
in case scaling is required for FB to be presented to default
panel.

To provide FB size option is provided via command line.
e.g
omapfb.fb_opt=0,720,1280,1,640,480
each element is [fb_ix,width,height]

Change-Id: I0b546932ff9119166e123b7bf08352ca836b0842
Signed-off-by: Dandawate Saket <dsaket@ti.com>