ARM: DRA7: hwmod data: add mmu data for ipu & dsp
A new MMU hwmod class and data structures are created to
represent the MMUs within the IPU and DSP processor subsystems
in DRA7xx. DRA7xx has two IPUs with an MMU each, and two DSPs
with two MMUs each, one for the processor core and the other
for the internal EDMA block.

NOTE:
1. The current irq information uses an implicit assignment of
   specific interrupt numbers for all MMUs (142 through 145) other
   than IPU1 MMU and DSP1 MMU0 (which use the default interrupt
   mappings). The actual interrupts are raised only when the
   corresponding crossbar registers are programmed correctly.
2. The irq information added in this patch should be removed
   for upstream.

Signed-off-by: Suman Anna <s-anna@ti.com>