OMAPDSS: HDMI: Fix GPA Channel Setting and Channel Allocation for 6-Channels
Enable channels 0 to 5 in GPA configuration (GP_CONF1) for the
6 channels case. This fixes an issue where the Audio Sample Packet
was carrying 8 channels instead of 6 channels.

The new GPA configuration sets the sample_present.sp3 bit to 0, then
0x13 is not a valid channel allocation. So, the channel allocation is
set to a valid value (0xb), when the number of channels is 6.

Also a typo is fixed in the comments for the eight channel case.

Change-Id: I3c987a4efdaf9f9594c99dc1fca1470b53a59355
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>