ASoC: davinci-mcasp: Add support for using McASP DATA port
By default the CFG port is used for data transfer. It
is desireable to use the DATA port so that the AFIFO
can be used on DRA7XX.

To enable DATA port usage, during the probe a check for
an IORESOURCE_MEM named "dat" is done. If the
resource is found, the data port will be used. If it is
not found, the driver will fall back to using the cfg
port.

Change-Id: I42c3cc04621354ef1fd7546f9feff003cf6b38ce
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>