regulator: core: Make set_voltage_tol() try for mid-range first
The expected semantic for something expressed as a tolerance is that it
should deliver the specified value with some deviation allowed but this
is not what set_voltage_tol() currently does. Instead it just passes
the maximum possible range to set_voltage() which will typically result
in a voltage aimed at lower than the target voltage.

Instead first try to set a voltage between the target voltage and the
upper limit, then fall back on the full range. This will be much more
robust against physical variation in systems and makes the API behave
more like users would expect.

Signed-off-by: Mark Brown <broonie@linaro.org>
[cherry-pick from
https://git.kernel.org/cgit/linux/kernel/git/broonie/regulator.git/commit/?h=topic/core&id=dc9ceed6a12aff627c81e01ada191e8a23fcbe3e
]
Change-Id: I52e88a886e4277e5f7d9f97192931290b9775eab
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>