OMAPDSS: HDMI: Fix GPA Channel Setting and Channel Allocation for 6-Channels
Enable channels 0 to 5 in GPA configuration (GP_CONF1) for the
6 channels case. This fixes an issue where the Audio Sample Packet
was carrying 8 channels instead of 6 channels.

The new GPA configuration sets the sample_present.sp3 bit to 0, then
0x13 is not a valid channel allocation. So, the channel allocation is
set to a valid value (0xb), when the number of channels is 6.

Also a typo is fixed in the comments for the eight channel case.

Change-Id: I6c9379db3977cc0d90f8e97112949a5e12322794
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>