ARM: DRA7: clock: initialize DSP DPLL and associated output clocks
The DSP DPLL is a new DPLL compared to previous OMAP generations and
is currently configured by bootloader. This DPLL supplies the root
clocks for the DSP processors, as well as a mux input source for EVE
sub-system. Initialize the DSP DPLL at boot time to avoid any functional
issues caused when any of the DSP processors. The DPLL will automatically
transition into a low-power stop mode when the associated output clocks
are not utilized or gated automatically.

The DSP DPLL provides two output clocks, DSP_GFCLK and EVE_GCLK. The
desired rate for DSP_GFCLK is 600 MHz (same as DSP DPLL CLKOUT frequency),
and is currently auto set due to the desired M2 divider value being the
same as reset value for the locked frequency of 600 MHz. The EVE_GCLK
however is required to be 400 MHz, so set the dpll_dsp_m3x2_ck's rate
explicitly so that the divider is set properly. The dpll_dsp_m2_ck rate
is also set explicitly to not rely on any implicit matching divider reset
values to the locked DPLL frequency.

The desired frequency values are defined as per the DRA7xx PLL spec
v0.3.5.

[a0783926@ti.com: DSP DPLL setting as per Rajendra Nayak's suggestion]
Signed-off-by: G Anthony <a0783926@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>